requirements.txt
pytorch
torchvision
cudatoolkit
pandas
numpy
matplotlib
tqdm
json


